requirements.txt
scikit-learn
mlflow
numpy
skl2onnx
gradio
langchain-openai
langchain
python-dotenv